#194 [REFACTOR] : postlist로 패키지명 수정
